Bioinformatics.jl Documentation

Bioinformatics.jl Documentation

Bioinformatics.Sequence — Type.

Structure for sequence data.

source

Bioinformatics.aliphatic_index — Method.

function aliphatic_index(x_ala, x_val, x_ile, x_leu::Float64)

The aliphatic index of a protein is defined as the relative volume occupied by aliphatic side chains (alanine, valine, isoleucine, and leucine).

See also: https://web.expasy.org/protparam/protparam-doc.html.

source

Bioinformatics.dotmatrix — Method.

function dotmatrix(s1::Sequence, s2::Sequence)

Calculate the dotplot matrix for given two sequences.

source

Bioinformatics.edit_dist — Method.

function edit_dist(s1::String, s2::String)

Calculate edit distance between two strings.

source

Bioinformatics.extinction_coeff — Method.

function extinction_coeff(n_tyr, n_trp, n_cys::Int64)

The extinction coefficient indicates how much light a protein absorbs at a certain wavelength. It is useful to have an estimation of this coefficient for following a protein which a spectrophotometer when purifying it.

See also: https://web.expasy.org/protparam/protparam-doc.html.

source

Bioinformatics.frequency — Method.

function frequency(seq::Sequence)

Count bases / amino acids for given sequence.

source

Bioinformatics.gc_content — Method.

function gc_content(seq::Sequence, window_size::Int64)

Calculate GC-content of given sequence using a window.

source

Bioinformatics.gc_content — Method.

function gc_content(seq::Sequence)

Calculate GC-content of given sequence.

source

Bioinformatics.gravy — Method.

function gravy(seq::Sequence)

The GRAVY value for a peptide or protein is calculated as the sum of hydropathy values of all the amino acids, divided by the number of residues in the sequence.

See also: https://web.expasy.org/protparam/protparam-doc.html.

source

Bioinformatics.hamming_dist — Method.

function hamming_dist(s1::String, s2::String)

Calculate Hamming distance between two strings.

source

Bioinformatics.instability_index — Method.

function instability_index(seq::Sequence)

The instability index provides an estimate of the stability of your protein in a test tube.

See also: https://web.expasy.org/protparam/protparam-doc.html.

source

Bioinformatics.plot_dotmatrix — Method.

function plot_dotmatrix(dotmatrix::Array{Int8})

Plot dotplot for given dotmatrix.

source

Bioinformatics.plot_gc_content — Method.

function plot_gc_content(seq::Sequence, window_size::Int64)

Plot GC-content of given sequence.

source

Bioinformatics.possible_proteins — Method.

function possible_proteins(s::Sequence)

Find possible proteins of given sequence.

source

Bioinformatics.protein_mass — Function.

function protein_mass(seq::Sequence, type = "monoisotopic")

Calculate mass of given amino acid sequence.

source

Bioinformatics.protparam — Method.

function protparam(seq::Sequence)

Computes various physico-chemical properties that can be deduced from a protein sequence.

See also: https://web.expasy.org/protparam/protparam-doc.html.

source

Bioinformatics.readFASTA — Method.

function readFASTA(filename::String)

Parse a FASTA formatted file.

source

Bioinformatics.reading_frames — Method.

function reading_frames(s::Sequence)

Find all reading frames of given sequence.

source

Bioinformatics.reverse_complement — Method.

function reverse_complement(s::Sequence)

Reverse complement of given DNA/RNA sequence.

source

Bioinformatics.transcription — Method.

function transcription(dna_seq::Sequence)

Transcribe DNA to RNA.

source

Bioinformatics.translation — Function.

function translation(s::Sequence, start_pos::Int64 = 1)

Translate given DNA sequence to Amino Acid sequence.

source

Bioinformatics.isoelectric_point — Method.

function isoelectric_point(seq::Sequence)

The isoelectric point, is the pH at which a molecule carries no net electrical charge or is electrically neutral in the statistical mean.

See also: https://web.expasy.org/protparam/protparam-doc.html, http://isoelectric.org/algorithms.html.

source

Bioinformatics.kmers — Function.

function kmers(s::Sequence, k::Int64 = 2)

Find all k-mers of given sequence.

source

Bioinformatics.kmers_frequency — Function.

function kmers_frequency(s::Sequence, k::Int64 = 2)

Calculate frequencies of k-mers of given sequence.

source